# refactored funcitons to use `sets.shift()` & `for .. of`