--- a/src/comparisons/equivalence.function.ts
+++ b/src/comparisons/equivalence.function.ts
@@ -26,9 +26,9 @@ export function equivalence<T, S extends ReadonlySet<T>>(...sets: S[]): boolean 
 	}
 
 	const primarySet = sets.shift()!;
-	for (const element of primarySet!) {
-		for (let index = 0; index < sets.length; ++index) {
-			if (!sets[index]!.has(element)) {
+	for (const element of primarySet) {
+		for (const set of sets) {
+			if (!set.has(element)) {
 				return false;
 			}
 		}