--- a/src/comparisons/proper-superset.function.ts
+++ b/src/comparisons/proper-superset.function.ts
@@ -25,9 +25,10 @@ export function properSuperset<T, S extends ReadonlySet<T>>(...sets: S[]): boole
 		return false;
 	}
 
-	for (let index = 1; index < sets.length; index++) {
+	const primarySet = sets.shift()!;
+	for (let index = 0; index < sets.length; ++index) {
 		for (const element of sets[index]!) {
-			if (!sets[0]?.has(element)) {
+			if (!primarySet.has(element)) {
 				return false;
 			}
 		}