--- a/src/comparisons/superset.function.ts
+++ b/src/comparisons/superset.function.ts
@@ -27,8 +27,8 @@ export function superset<T, S extends ReadonlySet<T>>(...sets: S[]): boolean {
 	}
 
 	const primarySet = sets.shift()!;
-	for (let index = 0; index < sets.length; ++index) {
-		for (const element of sets[index]!) {
+	for (const set of sets) {
+		for (const element of set) {
 			if (!primarySet.has(element)) {
 				return false;
 			}