--- a/src/operations/difference.function.ts
+++ b/src/operations/difference.function.ts
@@ -11,13 +11,13 @@ export function difference<T>(...sets: ReadonlySet<T>[]): ReadonlySet<T>;
  * @description A ∖ B ≔ { x : (x ∈ A) ∧ (x ∉ B) }
  */
 export function difference<T, S extends ReadonlySet<T>>(...sets: S[]): S {
-	const result = new Set<T>(sets[0]);
+	const resultSet = new Set<T>(sets.shift());
 
-	for (let index = 1; index < sets.length; index++) {
+	for (let index = 0; index < sets.length; ++index) {
 		for (const element of sets[index]!) {
-			result.delete(element);
+			resultSet.delete(element);
 		}
 	}
 
-	return result as ReadonlySet<T> as S;
+	return resultSet as ReadonlySet<T> as S;
 }