--- a/src/operations/difference.function.ts
+++ b/src/operations/difference.function.ts
@@ -13,8 +13,8 @@ export function difference<T>(...sets: ReadonlySet<T>[]): ReadonlySet<T>;
 export function difference<T, S extends ReadonlySet<T>>(...sets: S[]): S {
 	const resultSet = new Set<T>(sets.shift());
 
-	for (let index = 0; index < sets.length; ++index) {
-		for (const element of sets[index]!) {
+	for (const set of sets) {
+		for (const element of set) {
 			resultSet.delete(element);
 		}
 	}