--- a/src/operations/union.function.ts
+++ b/src/operations/union.function.ts
@@ -9,13 +9,13 @@ export function union<T>(...sets: ReadonlySet<T>[]): ReadonlySet<T>;
  * @description A ∪ B ≔ { x : (x ∈ A) ∨ (x ∈ B) }
  */
 export function union<T, S extends ReadonlySet<T>>(...sets: S[]): S {
-	const result = new Set<T>(sets[0]);
+	const resultSet = new Set<T>(sets.shift());
 
-	for (let index = 1; index < sets.length; index++) {
+	for (let index = 0; index < sets.length; ++index) {
 		for (const element of sets[index]!) {
-			result.add(element);
+			resultSet.add(element);
 		}
 	}
 
-	return result as ReadonlySet<T> as S;
+	return resultSet as ReadonlySet<T> as S;
 }