## Legibility
I personally find that the resulting code is more legible, using `sets.shift()` in conjunction with `for (const set of sets)`.
As an aside, this code style also removed all necessary usages of nullish coalescing.

## Performance
Moreover, the refactored code resulted in an overall performance improvement of ~11% overall. This is generally due to the fact that we are not fetching `sets[index]` in each loop iteration, in favor of a pre-cached `const set`.

Testing Notes:
- All tests were run on a 2017 Razer Blade Pro, Core i7-7700HQ Processor, 64GB ram.
- All tests were run using the V8 JS engine.
- All timings is in ms

<details>
<summary><h3 style="display: inline">[original] raw data:</h3></summary>

| Original Overall Timing: | time (ms)   |
|--------------------------|-------------|
| Original Run Timing 1    | 306,064.716 |
| Original Run Timing 2    | 320,495.549 |
| Original Run Timing 3    | 320,353.267 |
| Original Run Timing 4    | 324,881.274 |
| Original Run Timing 5    | 298,814.159 |
| Original Average         | 314,121.793 |


| Original Test Timing:                  | function(of1) | function(of1 of1) | function(of1 of2) | function(of2 of1) | function(of1 of1 of1) | function(of1 of2 of3) | function(of3 of2 of1) | function(...someEquivalent) | function(...manyEquivalent) | function(...someDisjoint) | function(...manyDisjoint) |
|----------------------------------------|---------------|-------------------|-------------------|-------------------|-----------------------|-----------------------|-----------------------|-----------------------------|-----------------------------|---------------------------|---------------------------|
|                                        |               |                   |                   |                   |                       |                       |                       |                             |                             |                           |                           |
| difference                             | 5,528.537     | 9,281.126         | 6,413.706         | 5,202.763         | 8,655.808             | 7,310.503             | 4,525.348             | 157.412                     | 147.759                     | 518.512                   | 326.498                   |
| difference                             | 5,165.204     | 9,319.900         | 6,352.209         | 5,225.966         | 9,312.768             | 7,924.933             | 5,175.987             | 161.351                     | 127.630                     | 564.634                   | 347.230                   |
| difference                             | 4,971.867     | 8,980.129         | 6,361.118         | 5,185.202         | 8,573.869             | 7,279.279             | 4,944.845             | 161.992                     | 157.714                     | 583.842                   | 333.448                   |
| difference                             | 5,086.717     | 9,639.039         | 6,953.099         | 5,565.336         | 9,386.295             | 7,694.673             | 4,949.082             | 183.770                     | 162.799                     | 648.022                   | 335.219                   |
| difference                             | 5,667.664     | 9,482.243         | 6,431.150         | 5,275.223         | 8,184.738             | 7,379.964             | 4,625.915             | 156.098                     | 136.462                     | 501.327                   | 308.818                   |
| original difference average            | 5,283.998     | 9,340.487         | 6,502.256         | 5,290.898         | 8,822.696             | 7,517.870             | 4,844.235             | 164.125                     | 146.473                     | 563.267                   | 330.243                   |
|                                        |               |                   |                   |                   |                       |                       |                       |                             |                             |                           |                           |
| intersection                           | 4,728.279     | 7,722.473         | 8,810.480         | 3,119.550         | 8,263.591             | 10,428.028            | 2,968.837             | 469.202                     | 283.654                     | 27.033                    | 0.257                     |
| intersection                           | 4,933.843     | 9,436.174         | 9,850.423         | 3,448.747         | 8,891.150             | 11,830.733            | 3,120.385             | 499.776                     | 299.540                     | 50.966                    | 0.450                     |
| intersection                           | 4,873.496     | 8,643.706         | 9,679.400         | 3,287.866         | 9,520.043             | 12,087.650            | 3,519.241             | 548.949                     | 307.196                     | 32.307                    | 0.322                     |
| intersection                           | 5,048.614     | 8,714.641         | 9,923.594         | 3,427.063         | 9,604.551             | 11,736.987            | 3,417.378             | 507.215                     | 306.311                     | 29.764                    | 0.786                     |
| intersection                           | 4,612.503     | 8,185.622         | 9,630.233         | 3,350.761         | 9,002.574             | 10,948.823            | 3,642.584             | 447.779                     | 290.996                     | 27.026                    | 0.257                     |
| original intersection average          | 4,839.347     | 8,540.523         | 9,578.826         | 3,326.797         | 9,056.382             | 11,406.444            | 3,333.685             | 494.584                     | 297.539                     | 33.419                    | 0.414                     |
|                                        |               |                   |                   |                   |                       |                       |                       |                             |                             |                           |                           |
| union                                  | 4,795.467     | 7,062.333         | 6,084.723         | 5,184.707         | 8,877.905             | 6,266.018             | 5,810.776             | 388.479                     | 241.289                     | 2,464.073                 | 2,638.573                 |
| union                                  | 5,117.724     | 7,671.835         | 6,240.681         | 5,487.634         | 9,501.192             | 6,812.853             | 5,753.113             | 532.984                     | 298.962                     | 2,470.075                 | 2,474.088                 |
| union                                  | 5,486.514     | 7,749.413         | 5,983.009         | 5,706.268         | 9,163.675             | 6,657.030             | 5,901.155             | 754.564                     | 252.291                     | 2,748.041                 | 2,968.163                 |
| union                                  | 5,016.722     | 7,625.566         | 6,112.399         | 5,710.951         | 9,713.799             | 7,103.469             | 5,922.020             | 498.116                     | 269.847                     | 2,721.073                 | 2,874.513                 |
| union                                  | 4,614.345     | 7,237.181         | 5,515.699         | 5,146.438         | 9,107.971             | 6,525.618             | 5,785.559             | 411.243                     | 242.027                     | 2,421.232                 | 2,797.015                 |
| original union average                 | 5,006.154     | 7,469.266         | 5,987.302         | 5,447.200         | 9,272.908             | 6,672.998             | 5,834.525             | 517.077                     | 260.883                     | 2,564.899                 | 2,750.470                 |
|                                        |               |                   |                   |                   |                       |                       |                       |                             |                             |                           |                           |
| xor                                    | 5,213.459     | 13,255.000        | 8,112.156         | 9,663.168         | 16,079.426            | 10,144.474            | 12,790.923            | 856.500                     | 363.123                     | 3,831.277                 | 3,535.291                 |
| xor                                    | 5,131.879     | 14,043.527        | 8,892.519         | 10,306.754        | 17,900.462            | 11,732.074            | 12,879.764            | 900.448                     | 402.814                     | 3,745.252                 | 3,691.311                 |
| xor                                    | 5,211.524     | 14,650.832        | 8,783.771         | 10,400.013        | 17,190.609            | 11,425.864            | 13,288.162            | 901.185                     | 435.490                     | 3,999.618                 | 3,832.765                 |
| xor                                    | 5,299.869     | 14,638.828        | 8,705.744         | 10,958.786        | 16,693.414            | 11,824.148            | 13,497.853            | 844.972                     | 381.220                     | 3,963.793                 | 3,811.337                 |
| xor                                    | 4,795.892     | 13,570.420        | 7,919.912         | 9,884.553         | 14,406.393            | 10,038.748            | 11,099.158            | 645.880                     | 360.885                     | 3,364.754                 | 3,346.572                 |
| original xor average                   | 5,130.525     | 14,031.721        | 8,482.820         | 10,242.655        | 16,454.061            | 11,033.062            | 12,711.172            | 829.797                     | 388.706                     | 3,780.939                 | 3,643.455                 |
|                                        |               |                   |                   |                   |                       |                       |                       |                             |                             |                           |                           |
| disjoint                               | 0.169         | 5,423.150         | 5,060.450         | 2,280.614         | 4,848.766             | 5,223.560             | 1,639.984             | 25.032                      | 0.143                       | 3,430.196                 | 3,744.891                 |
| disjoint                               | 0.107         | 5,288.349         | 5,477.972         | 2,387.449         | 5,007.689             | 5,122.024             | 1,633.462             | 13.834                      | 0.091                       | 3,491.174                 | 3,788.521                 |
| disjoint                               | 0.222         | 5,605.339         | 5,367.305         | 2,463.414         | 5,600.900             | 5,989.221             | 1,662.269             | 13.797                      | 0.090                       | 3,598.689                 | 3,877.460                 |
| disjoint                               | 0.117         | 5,507.348         | 5,350.934         | 2,475.010         | 5,101.607             | 5,261.371             | 1,699.821             | 12.881                      | 0.088                       | 3,640.761                 | 3,982.408                 |
| disjoint                               | 0.091         | 4,679.158         | 4,582.319         | 2,135.173         | 4,500.166             | 4,961.522             | 1,551.858             | 12.761                      | 0.115                       | 3,235.288                 | 3,730.767                 |
| original disjoint average              | 0.141         | 5,300.669         | 5,167.796         | 2,348.332         | 5,011.826             | 5,311.540             | 1,637.479             | 15.661                      | 0.105                       | 3,479.222                 | 3,824.809                 |
|                                        |               |                   |                   |                   |                       |                       |                       |                             |                             |                           |                           |
| equivalence                            | 0.102         | 3,765.865         | 4.947             | 0.024             | 4,258.086             | 5.076                 | 0.022                 | 1,375.723                   | 2,184.956                   | 0.087                     | 3.236                     |
| equivalence                            | 0.196         | 3,487.337         | 4.631             | 0.039             | 3,935.162             | 6.110                 | 0.023                 | 1,299.788                   | 1,932.168                   | 0.091                     | 3.084                     |
| equivalence                            | 0.134         | 3,793.092         | 4.815             | 0.046             | 4,030.678             | 5.335                 | 0.030                 | 1,276.185                   | 1,863.609                   | 0.146                     | 1.872                     |
| equivalence                            | 0.123         | 3,577.098         | 5.812             | 0.036             | 4,083.099             | 6.060                 | 0.022                 | 1,258.498                   | 1,999.499                   | 0.073                     | 1.776                     |
| equivalence                            | 0.112         | 3,274.627         | 5.011             | 0.026             | 3,692.570             | 5.123                 | 0.020                 | 1,173.084                   | 1,763.635                   | 0.093                     | 1.980                     |
| original equivalence average           | 0.133         | 3,579.604         | 5.043             | 0.034             | 3,999.919             | 5.541                 | 0.023                 | 1,276.656                   | 1,948.773                   | 0.098                     | 2.390                     |
|                                        |               |                   |                   |                   |                       |                       |                       |                             |                             |                           |                           |
| properSubset                           | 0.166         | 0.080             | 0.024             | 1,871.919         | 5.412                 | 0.028                 | 0.028                 | 0.106                       | 1.807                       | 0.057                     | 1.469                     |
| properSubset                           | 0.117         | 0.058             | 0.038             | 1,785.750         | 5.948                 | 0.027                 | 0.024                 | 0.055                       | 1.649                       | 0.056                     | 1.881                     |
| properSubset                           | 0.102         | 0.040             | 0.018             | 1,848.702         | 5.757                 | 0.027                 | 0.025                 | 0.054                       | 1.687                       | 0.068                     | 1.495                     |
| properSubset                           | 0.131         | 0.049             | 0.048             | 2,008.718         | 6.311                 | 0.026                 | 0.023                 | 0.050                       | 1.719                       | 0.066                     | 2.147                     |
| properSubset                           | 0.149         | 0.061             | 0.033             | 1,986.183         | 6.484                 | 0.026                 | 0.024                 | 0.054                       | 1.506                       | 0.068                     | 1.989                     |
| original properSubset average          | 0.133         | 0.058             | 0.032             | 1,900.254         | 5.982                 | 0.027                 | 0.025                 | 0.064                       | 1.674                       | 0.063                     | 1.796                     |
|                                        |               |                   |                   |                   |                       |                       |                       |                             |                             |                           |                           |
| properSuperset                         | 0.105         | 0.043             | 1,990.500         | 5.095             | 0.025                 | 2,471.807             | 4.959                 | 0.067                       | 0.822                       | 0.035                     | 0.678                     |
| properSuperset                         | 0.103         | 0.039             | 1,742.347         | 6.472             | 0.030                 | 2,528.196             | 5.643                 | 0.041                       | 0.777                       | 0.036                     | 0.960                     |
| properSuperset                         | 0.102         | 0.040             | 1,720.644         | 5.737             | 0.027                 | 2,531.367             | 4.848                 | 0.059                       | 0.853                       | 0.037                     | 0.640                     |
| properSuperset                         | 0.104         | 0.046             | 1,820.529         | 7.214             | 0.043                 | 2,573.628             | 5.389                 | 0.039                       | 0.777                       | 0.031                     | 0.659                     |
| properSuperset                         | 0.120         | 0.042             | 1,739.563         | 4.770             | 0.042                 | 2,642.126             | 6.754                 | 0.072                       | 0.887                       | 0.061                     | 0.671                     |
| original properSuperset average        | 0.107         | 0.042             | 1,802.717         | 5.858             | 0.033                 | 2,549.425             | 5.519                 | 0.056                       | 0.823                       | 0.040                     | 0.722                     |
|                                        |               |                   |                   |                   |                       |                       |                       |                             |                             |                           |                           |
| subset                                 | 0.124         | 3,492.658         | 5.927             | 1,832.577         | 3,564.048             | 0.021                 | 0.057                 | 1,275.574                   | 1,810.200                   | 0.041                     | 0.709                     |
| subset                                 | 0.107         | 3,534.076         | 7.095             | 1,773.470         | 3,727.766             | 0.021                 | 0.057                 | 1,328.451                   | 1,995.200                   | 0.057                     | 0.821                     |
…